Separate synchronous implementations
