+++ driver/accel/synchronous.h
@@ -3,23 +3,11 @@
 
 #include "../fixedptc.h"
 
-#define SYNCHRONOUS_GAIN_LUT_START -3
-#define SYNCHRONOUS_GAIN_LUT_STOP 9
-#define SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE 8
-#define SYNCHRONOUS_GAIN_LUT_SIZE                                              \
-  ((SYNCHRONOUS_GAIN_LUT_STOP - SYNCHRONOUS_GAIN_LUT_START) *                  \
-       SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE +                                \
-   1)
-
 struct synchronous_curve_args {
   fpt gamma;
   fpt smooth;
   fpt motivity;
   fpt sync_speed;
-};
-
-struct synchronous_gain_lut {
-  fpt velocity[SYNCHRONOUS_GAIN_LUT_SIZE];
 };
 
 static inline fpt __synchronous_ln(fpt value) {
@@ -100,78 +88,5 @@ static inline fpt __synchronous_sens_fun(fpt input_speed,
                             sharpness_recip);
     return fpt_exp(fpt_mul(exponent, log_motivity));
   }
-}
-
-static inline fpt __synchronous_gain_pow2(int exponent) {
-  if (exponent >= 0)
-    return fpt_fromint(1 << exponent);
-
-  return FIXEDPT_ONE >> -exponent;
-}
-
-static inline fpt __synchronous_gain_x(int index) {
-  int octave = index / SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE;
-  int step = index % SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE;
-  fpt octave_start =
-      __synchronous_gain_pow2(SYNCHRONOUS_GAIN_LUT_START + octave);
-
-  return fpt_mul(fpt_fromint(step + SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE),
-                 fpt_div(octave_start,
-                         fpt_fromint(SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE)));
-}
-
-static inline void
-__synchronous_gain_lut_init(struct synchronous_gain_lut *lut,
-                            struct synchronous_curve_args args) {
-  fpt sum = 0;
-  fpt previous_x = 0;
-  int i;
-
-  for (i = 0; i < SYNCHRONOUS_GAIN_LUT_SIZE; i++) {
-    fpt x = __synchronous_gain_x(i);
-    fpt interval = fpt_div(x - previous_x, fpt_fromint(2));
-
-    sum +=
-        fpt_mul(__synchronous_sens_fun(previous_x + interval, args), interval);
-    sum += fpt_mul(__synchronous_sens_fun(x, args), interval);
-    lut->velocity[i] = sum;
-    previous_x = x;
-  }
-}
-
-static inline fpt __synchronous_gain_lerp(fpt a, fpt b, fpt t) {
-  return a + fpt_mul(t, b - a);
-}
-
-static inline fpt
-__synchronous_gain_lut_lookup(fpt input_speed,
-                              const struct synchronous_gain_lut *lut) {
-  const fpt x_start = __synchronous_gain_pow2(SYNCHRONOUS_GAIN_LUT_START);
-  int exponent = SYNCHRONOUS_GAIN_LUT_START;
-  fpt base;
-  fpt index_f;
-  int index;
-  fpt velocity;
-
-  if (input_speed <= x_start)
-    return fpt_div(lut->velocity[0], x_start);
-
-  while (exponent < SYNCHRONOUS_GAIN_LUT_STOP - 1 &&
-         input_speed >= __synchronous_gain_pow2(exponent + 1))
-    exponent++;
-
-  base = __synchronous_gain_pow2(exponent);
-  index_f = fpt_fromint((exponent - SYNCHRONOUS_GAIN_LUT_START) *
-                        SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE) +
-            fpt_mul(fpt_div(input_speed - base, base),
-                    fpt_fromint(SYNCHRONOUS_GAIN_LUT_POINTS_PER_OCTAVE));
-  index = fpt_toint(index_f);
-  if (index > SYNCHRONOUS_GAIN_LUT_SIZE - 2)
-    index = SYNCHRONOUS_GAIN_LUT_SIZE - 2;
-
-  velocity =
-      __synchronous_gain_lerp(lut->velocity[index], lut->velocity[index + 1],
-                              index_f - fpt_fromint(index));
-  return fpt_div(velocity, input_speed);
 }
 #endif
